fsck: Add the capability to fix zero size directory
Add the capability to fix the directory which size=0 and starts with
0 cluster. This kind of directory often made by Paragon ufsd dirver,
and it will cause failure when we try to it using linux exfat driver.
The way to fix it, we should find a cluster and assign it to
the directory's first cluster and modify the size to 1 (Like exfat
driver does during mkdir).

Signed-off-by: Erichong <erichong@qnap.com>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>